+++ css/mainStyle.css
@@ -19,127 +19,28 @@ body {
 }
 
 /* Header */
-nav {
-  position: fixed;
-  z-index: 999;
-  color: var(--red-99);
+header {
   width: 100%;
   height: 80px;
   display: flex;
   flex-direction: row;
   align-items: center;
-  justify-content: space-around;
-  background-color: var(--black-10);
-  padding: 50px 0;
-}
-
-#logo-section {
-  display: flex;
-}
-
-#logo-section p {
-  cursor: pointer;
-  font-size: 1.6rem;
-  padding-right: 20px;
-}
-
-.mypage {
-  padding: 0 10px;
-  margin-right: 10px;
-  border-right: 1px solid var(--black-90);
 }
 
 #header-logo {
-  width: 250px;
+  width: 170px;
   height: auto;
-  margin-right: 30px;
+  position: fixed;
+  margin-left: 20px;
   z-index: 1;
   cursor: pointer;
-}
-
-#login-section {
-  margin-left: 150px;
-  display: flex;
-  align-items: center;
-  justify-content: end;
-}
-
-.login-button {
-  background-color: var(--red-20);
-  border: 1px solid var(--red-30);
-  border-radius: 3px;
-  cursor: pointer;
-  color: var(--red-99);
-  padding: 10px 15px;
-  font-size: 1.6rem;
-  margin-left: 20px;
-}
-
-.login-button:hover,
-.login-button:active {
-  background-color: initial;
-  color: var(--red-30);
-}
-
-/* toggle */
-fieldset {
-  border: none;
-  font-size: 1.3rem;
-}
-
-label {
-  display: inline-flex;
-  align-items: center;
-  gap: 0.7rem;
-  cursor: pointer;
-}
-
-#toggle,
-#language-toggle {
-  appearance: none;
-  position: relative;
-  border: max(2px, 0.1em) solid var(--black-99);
-  border-radius: 1.25em;
-  width: 2.25em;
-  height: 1.25em;
-}
-
-#toggle::before,
-#language-toggle::before {
-  content: "";
-  position: absolute;
-  left: 0;
-  width: 1em;
-  height: 1em;
-  border-radius: 50%;
-  transform: scale(0.8);
-  background-color: var(--black-99);
-  transition: left 250ms linear;
-}
-
-#toggle:checked,
-#language-toggle:checked {
-  background-color: var(--red-30);
-  border-color: var(--red-20);
-}
-
-#toggle:enabled:hover,
-#language-toggle:enabled:hover {
-  box-shadow: 0 0 0 max(1px, 0.1em) var(--black-99);
-  cursor: pointer;
-}
-
-#toggle:checked::before,
-#language-toggle:checked::before {
-  background-color: var(--black-99);
-  left: 1em;
 }
 
 #search-box {
   display: flex;
   justify-content: center;
   align-items: center;
-  padding-top: 130px;
+  margin: 0 auto;
 }
 
 #search-input {
@@ -166,32 +67,6 @@ label {
   width: 20px;
   height: auto;
   cursor: pointer;
-}
-
-/* pagination-section */
-#pagination-section {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  margin-bottom: 70px;
-}
-
-.page-number-btn,
-#prev-page,
-#next-page {
-  cursor: pointer;
-  color: var(--black-99);
-  background-color: transparent;
-  border: none;
-  font-size: 1.9rem;
-}
-
-.page-number-btn:hover {
-  color: var(--black-40);
-}
-
-.page-number-btn.active {
-  color: var(--red-30);
 }
 
 /* movie-card */
