+++ index.html
@@ -9,17 +9,15 @@
     <link rel="stylesheet" href="./css/mainStyle.css" />
   </head>
   <body>
-    <form id="search-box">
-      <input type="text" id="search-input" placeholder="원하는 영화 제목을 입력하세요" />
-      <button id="search-button">검색</button>
-    </form>
+    <header>
+      <img src="./img/logo_WHITE.png" alt="logo" id="header-logo" />
+      <form id="search-box">
+        <input type="text" id="search-input" placeholder="원하는 영화 제목을 입력하세요" />
+        <button id="search-button">검색</button>
+      </form>
+    </header>
     <section id="content-area">
       <ul id="card-container"></ul>
     </section>
-    <div id="pagination-section">
-      <button id="prev-page"><</button>
-      <span id="page-number">1</span>
-      <button id="next-page">></button>
-    </div>
   </body>
 </html>
